% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensig.R
\name{gensig}
\alias{gensig}
\title{Simulating genetic signatures of disease transmission}
\usage{
gensig(param = NULL, config = NULL)
}
\arguments{
\item{param}{a list of lists describing the parameter values for each pathogen of interest}

\item{config}{an optional list of parameter values for the run}
}
\value{
Returns a dataframe of two columns
}
\description{
Simulating genetic signatures of disease transmission
}
\author{
Finlay Campbell <f.campbell15@imperial.ac.uk>
}

