% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dist}
\alias{plot.dist}
\title{Visualise distributions of genetic signatures}
\usage{
\method{plot}{dist}(store, config = create.config())
}
\description{
Visualise distributions of genetic signatures
}

