% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vis.dist}
\alias{vis.dist}
\title{Visualise distributions of genetic signatures}
\usage{
vis.dist(store, config = create.config())
}
\arguments{
\item{store}{A gensig output dataframe}

\item{config}{A gensig config list}
}
\description{
Visualise distributions of genetic signatures
}

