% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vis.mean}
\alias{vis.mean}
\title{Compare mean genetic signatures between pathogens}
\usage{
vis.mean(store)
}
\arguments{
\item{store}{A gensig output dataframe}
}
\description{
Compare mean genetic signatures between pathogens
}

