% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vis.w}
\alias{vis.w}
\title{Visualise distributions of generation times}
\usage{
vis.w(store, param = create.param())
}
\arguments{
\item{store}{A gensig output dataframe}

\item{param}{A gensig param list}
}
\description{
Visualise distributions of generation times
}

